Include the required CSS and JS inline
The CSS and JS wasn't loading correctly for me. I think there's a bug in the $path code. I think a better approach would be to just inline the CSS and JS so people don't have to worry about how to serve those files publicly when their Zebra code is probably stored away in a private directory.