realpath for auto resource_path
Recently switched to a symlinked doc root and noticed an interesting problem, `dirname(__FILE__)` resolves to the real resolved path (not the symlink) and `$_SERVER['DOCUMENT_ROOT']` resolves to the unresolved symlink path. So if your paths don't have the same amount of characters it causes a problem.
The `DIRECTORY_SEPARATOR` doesn't seem to be needed, at least on my dev server it adds an extra / between domain and path but it doesn't do anything bad so I figured I'd leave it in case its needed elsewhere.